fix guestOsMapper and move mapping code to latest upgrade
